phase 2 - mousePos etc
